+++ src/js/views/dashboard/widget/table/wizard/index.js
@@ -8,21 +8,10 @@ import { v4 as uuidv4 } from 'uuid';
 import useTable from '../../wizard/hooks/useTable';
 import Attributes from '../../wizard/Steps/Attributes';
 import Devices from '../../wizard/Steps/Devices';
-import { GeneralFilter } from '../../wizard/Steps/Filters';
-import General from '../../wizard/Steps/General';
+import Filters from '../../wizard/Steps/Filters';
+import General, { generalValidates } from '../../wizard/Steps/General';
 import Summary from '../../wizard/Steps/Summary';
 import Wizard from '../../wizard/wizard';
-
-const generalValidates = values => {
-  const errors = {};
-  if (!values.general.name) {
-    errors.name = 'Required';
-  }
-  if (!values.general.description) {
-    errors.description = 'Required';
-  }
-  return errors;
-};
 
 const stepsList = [
   { label: 'steps.general', key: uuidv4() },
